+++ app/Http/Controllers/API/v1/PrivateMangaReadlistController.php
@@ -11,6 +11,7 @@ use App\Http\Response\CustomResponse;
 use Illuminate\Http\Request;
 use App\Http\Resources\v1\ReadlistResource;
 use App\Http\Resources\v1\PrivateMangaReadlistCollection;
+use App\Models\PrivateMangaReadlist;
 
 class PrivateMangaReadlistController extends Controller
 {
@@ -26,6 +27,7 @@ class PrivateMangaReadlistController extends Controller
      */
     public function index(User $user, Readlist $readlist, Request $request)
     {
+        $this->authorize('viewAny', [PrivateMangaReadlist::class, $readlist]);
         $readlist->load('privateMangas');
         $paginatedReadlistMangas = $readlist->privateMangas()->paginate(10);
 
@@ -42,6 +44,7 @@ class PrivateMangaReadlistController extends Controller
         User $user,
         Readlist $readlist)
     {
+        $this->authorize('create', [PrivateMangaReadlist::class, $readlist]);
         $readlist->privateMangas()->attach($request->manga_id);
 
         return $this->customResponse->createdResponse();
@@ -52,6 +55,7 @@ class PrivateMangaReadlistController extends Controller
      */
     public function destroy(User $user, Readlist $readlist, PrivateManga $privateManga)
     {
+        $this->authorize('delete', [PrivateMangaReadlist::class, $readlist]);
         $readlist->privateMangas()->detach($privateManga);
 
         return $this->customResponse->deletedResponse();
